refactor(sync_handler/v1): seperate conv to follow v0 convention
