+++ crates/pathfinder/src/p2p_network/conv.rs
@@ -1,9 +1,6 @@
 //! Workaround for the orphan rule - implement conversion fns for types ourside our crate.
 
-use std::{
-    collections::HashMap,
-    time::{Duration, SystemTime},
-};
+use std::{collections::HashMap, time::SystemTime};
 
 use pathfinder_common::{
     state_update::{ContractClassUpdate, ContractUpdate, SystemContractUpdate},
@@ -12,103 +9,11 @@ use pathfinder_common::{
     StateCommitment, StateUpdate, StorageAddress, StorageCommitment, StorageValue,
     TransactionCommitment,
 };
-use stark_hash::Felt;
-
-pub trait ToProto<T> {
-    fn to_proto(self) -> T;
-}
 
 pub trait TryFromProto<T> {
     fn try_from_proto(proto: T) -> anyhow::Result<Self>
     where
         Self: Sized;
-}
-
-impl ToProto<p2p_proto_v1::block::BlockHeader> for BlockHeader {
-    fn to_proto(self) -> p2p_proto_v1::block::BlockHeader {
-        use p2p_proto_v1::common::{Address, ChainId, Hash, Merkle};
-        const ZERO_MERKLE: Merkle = Merkle {
-            n_leaves: 0,
-            root: Hash(Felt::ZERO),
-        };
-        p2p_proto_v1::block::BlockHeader {
-            parent_block: Hash(self.parent_hash.0),
-            time: SystemTime::UNIX_EPOCH // FIXME Dunno how to convert
-                .checked_add(Duration::from_secs(self.timestamp.get()))
-                .unwrap(),
-            sequencer_address: Address(self.sequencer_address.0),
-            // FIXME: all of those zeros
-            state_diffs: ZERO_MERKLE,
-            state: ZERO_MERKLE,
-            proof_fact: Hash(Felt::ZERO),
-            transactions: ZERO_MERKLE,
-            events: ZERO_MERKLE,
-            receipts: ZERO_MERKLE,
-            protocol_version: 0,
-            chain_id: ChainId(Felt::ZERO),
-            // FIXME extra fields added to make sync work
-            block_hash: Hash(self.hash.0),
-            gas_price: self.gas_price.0.to_be_bytes().into(),
-            starknet_version: self.starknet_version.take_inner(),
-        }
-    }
-}
-
-impl ToProto<p2p_proto_v1::state::StateDiff> for StateUpdate {
-    fn to_proto(self) -> p2p_proto_v1::state::StateDiff {
-        use p2p_proto_v1::common::Address;
-        use p2p_proto_v1::state::{ContractDiff, ContractStoredValue, StateDiff};
-        StateDiff {
-            tree_id: 0, // TODO there will initially be 2 trees, dunno which id is which
-            contract_diffs: self
-                .system_contract_updates
-                .into_iter()
-                .map(|(address, update)| {
-                    let address = Address(address.0);
-                    let values = update
-                        .storage
-                        .into_iter()
-                        .map(|(storage_address, storage_value)| ContractStoredValue {
-                            key: storage_address.0,
-                            value: storage_value.0,
-                        })
-                        .collect();
-                    let class_hash = Felt::ZERO;
-                    let nonce = Felt::ZERO; // FIXME cannot distinguish between None and real 0 nonce
-                    ContractDiff {
-                        address,
-                        nonce,
-                        class_hash,
-                        values,
-                    }
-                })
-                .chain(self.contract_updates.into_iter().map(|(address, update)| {
-                    let address = Address(address.0);
-                    let ContractUpdate {
-                        storage,
-                        class,
-                        nonce,
-                    } = update;
-                    let values = storage
-                        .into_iter()
-                        .map(|(storage_address, storage_value)| ContractStoredValue {
-                            key: storage_address.0,
-                            value: storage_value.0,
-                        })
-                        .collect();
-                    let class_hash = class.map(|c| c.class_hash()).unwrap_or_default().0;
-                    let nonce = nonce.unwrap_or_default().0;
-                    ContractDiff {
-                        address,
-                        nonce,
-                        class_hash,
-                        values,
-                    }
-                }))
-                .collect(),
-            // FIXME missing: declared classes cairo & sierra, replaced classes, old and new state commitments
-        }
-    }
 }
 
 // FIXME at the moment this implementation is useless due to the massive difference between p2p and internal header representations,
